Add abstraction for Android's vibrator service
