+++ app/ui/legacy/src/main/java/com/fsck/k9/ui/settings/KoinModule.kt
@@ -3,6 +3,7 @@ package com.fsck.k9.ui.settings
 import com.fsck.k9.helper.NamedThreadFactory
 import com.fsck.k9.ui.settings.account.AccountSettingsDataStoreFactory
 import com.fsck.k9.ui.settings.account.AccountSettingsViewModel
+import com.fsck.k9.ui.settings.account.getSystemVibrator
 import com.fsck.k9.ui.settings.export.SettingsExportViewModel
 import com.fsck.k9.ui.settings.general.GeneralSettingsDataStore
 import com.fsck.k9.ui.settings.general.GeneralSettingsViewModel
@@ -33,6 +34,7 @@ val settingsUiModule = module {
             notificationController = get()
         )
     }
+    factory { getSystemVibrator(context = get()) }
 
     viewModel { SettingsExportViewModel(context = get(), preferences = get(), settingsExporter = get()) }
     viewModel { SettingsImportViewModel(get(), get()) }
